move get daily balances logic into a cron
also refactor logic into dedicatd tbc_accounts.ts file and switch up
the params a bit to be less restrictive (so we can fetch different dates
for different accounts, e.g. if one was recently added or something)
