+++ package.json
@@ -20,12 +20,14 @@
     "@solana/web3.js": "^1.41.4",
     "@types/express": "^4.17.13",
     "@types/node": "^17.0.31",
+    "@types/node-cron": "^3.0.1",
     "bs58": "^5.0.0",
     "cross-fetch": "^3.1.5",
     "dotenv": "^16.0.0",
     "express": "^4.18.1",
     "graphql-request": "^4.2.0",
     "knex": "^2.0.0",
+    "node-cron": "^3.0.0",
     "pg": "^8.7.3",
     "ts-node": "^10.7.0",
     "typescript": "^4.6.4"
