+++ src/app.ts
@@ -1,6 +1,6 @@
 import express from "express";
-
 import routes from "./routes";
+import { initCron } from "./cron";
 
 class App {
   public server;
@@ -10,6 +10,8 @@ class App {
 
     this.middlewares();
     this.routes();
+
+    initCron();
   }
 
   middlewares() {
